Move Shared module
